Reorganize frontend authoring files
